Refactor exceptions, add new exceptions, make batch exception final, make all exceptions immutable
