+++ src/Exception/ClientException.php
@@ -1,0 +1,13 @@
+<?php
+
+namespace Http\Client\Exception;
+
+/**
+ * Thrown when a client error (4xx) is encountered
+ *
+ * @author Márk Sági-Kazár <mark.sagikazar@gmail.com>
+ */
+class ClientException extends HttpException
+{
+
+}
