+++ src/Exception/NetworkException.php
@@ -1,0 +1,13 @@
+<?php
+
+namespace Http\Client\Exception;
+
+/**
+ * Thrown when the request cannot be completed caused by network issues
+ *
+ * @author Márk Sági-Kazár <mark.sagikazar@gmail.com>
+ */
+class NetworkException extends RequestException
+{
+
+}
