+++ src/Exception/ServerException.php
@@ -1,0 +1,13 @@
+<?php
+
+namespace Http\Client\Exception;
+
+/**
+ * Thrown when a server error (5xx) is encountered
+ *
+ * @author Márk Sági-Kazár <mark.sagikazar@gmail.com>
+ */
+class ServerException extends HttpException
+{
+
+}
