+++ src/Exception/TransferException.php
@@ -1,0 +1,13 @@
+<?php
+
+namespace Http\Client\Exception;
+
+/**
+ * Thrown when an error related to sending/receiving the request/response occurs
+ *
+ * @author Márk Sági-Kazár <mark.sagikazar@gmail.com>
+ */
+class TransferException extends RuntimeException
+{
+
+}
